requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0 # Downgraded from 2.10.1 to resolve conflict with djangorestframework_simplejwt
sqlparse==0.5.3
tzdata==2025.2
